Fix for bug #756
The problem was that the mouse capture was not being reset
when the pointer left the track panel.  In that case, as
long as no buttons are down, the capture state needs to be
reset so that when the mouse re-enters the track panel it
figures out what is going on from a clean slate.